- Add -Wextra -Wcast-align -Wbad-function-cast to CFLAGS if --enable-warnings is specified - Fixed several compile warnings - 64-bit cleanliness fixes, e.g., reorganize data structures to reduce storage/unnecessary padding
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid-7.2@1013 82007160-df01-0410-b94d-b575c5fd34c7

+++ include/list.h
@@ -65,7 +65,7 @@ struct _dlink_list
 {
   dlink_node *head;
   dlink_node *tail;
-  unsigned long length;
+  unsigned int length;
 };
 
 extern void dlinkAdd(void *, dlink_node *, dlink_list *);
@@ -79,5 +79,4 @@ extern dlink_node *dlinkFindDelete(dlink_list *, void *);
 extern void init_dlink_nodes(void);
 extern void free_dlink_node(dlink_node *);
 extern dlink_node *make_dlink_node(void);
-
 #endif
